refactor: 抽元件 userDropdownMenu
